[Fresh] Generate signatures for Hooks (#15733)
* Generate signatures for Hooks

This currently only works one level deep. For custom Hooks, we'll need to add some way to compose signatures.

* Be more resilient to plugin conflicts

This prevents a class of problems where other plugins cause our visitor to re-run.

It's a standard Babel practice, e.g.:

https://github.com/babel/babel/blob/8c7d4b55c99ff34cb9d493d452472e59b5ed1e70/packages/babel-plugin-transform-react-constant-elements/src/index.js#L85-L86

* Remove unnecessary stuff from debugging

* Include Foo.useHookName() calls into the signature

* Add an integration test for adding/removing an effect

* Add integration test for changing custom Hook order

* Include custom Hooks into the signatures

* Fix inferred names for function expressions

* Support export default hoc(Foo) when Foo is defined separately

* Add more built-in Hooks

+++ packages/react-fresh/src/ReactFreshBabelPlugin.js
@@ -8,7 +8,7 @@
 'use strict';
 
 export default function(babel) {
-  const {types: t, template} = babel;
+  const {types: t} = babel;
 
   const registrationsByProgramPath = new Map();
   function createRegistration(programPath, persistentID) {
@@ -24,10 +24,6 @@ export default function(babel) {
     return handle;
   }
 
-  const buildRegistrationCall = template(`
-    __register__(HANDLE, PERSISTENT_ID);
-  `);
-
   function isComponentishName(name) {
     return typeof name === 'string' && name[0] >= 'A' && name[0] <= 'Z';
   }
@@ -35,6 +31,15 @@ export default function(babel) {
   function findInnerComponents(inferredName, path, callback) {
     const node = path.node;
     switch (node.type) {
+      case 'Identifier': {
+        if (!isComponentishName(node.name)) {
+          return false;
+        }
+        // export default hoc(Foo)
+        // const X = hoc(Foo)
+        callback(inferredName, node, null);
+        return true;
+      }
       case 'FunctionDeclaration': {
         // function Foo() {}
         // export function Foo() {}
@@ -158,6 +163,80 @@ export default function(babel) {
     return false;
   }
 
+  let hookCalls = new WeakMap();
+
+  function recordHookCall(functionNode, hookCallPath, hookName) {
+    if (!hookCalls.has(functionNode)) {
+      hookCalls.set(functionNode, []);
+    }
+    let hookCallsForFn = hookCalls.get(functionNode);
+    let key = '';
+    if (hookCallPath.parent.type === 'VariableDeclarator') {
+      // TODO: if there is no LHS, consider some other heuristic.
+      key = hookCallPath.parentPath.get('id').getSource();
+    }
+    hookCallsForFn.push({
+      name: hookName,
+      callee: hookCallPath.node.callee,
+      key,
+    });
+  }
+
+  function isBuiltinHook(hookName) {
+    switch (hookName) {
+      case 'useState':
+      case 'React.useState':
+      case 'useReducer':
+      case 'React.useReducer':
+      case 'useEffect':
+      case 'React.useEffect':
+      case 'useLayoutEffect':
+      case 'React.useLayoutEffect':
+      case 'useMemo':
+      case 'React.useMemo':
+      case 'useCallback':
+      case 'React.useCallback':
+      case 'useRef':
+      case 'React.useRef':
+      case 'useContext':
+      case 'React.useContext':
+      case 'useImperativeMethods':
+      case 'React.useImperativeMethods':
+      case 'useDebugValue':
+      case 'React.useDebugValue':
+        return true;
+      default:
+        return false;
+    }
+  }
+
+  function getHookCallsSignature(functionNode) {
+    const fnHookCalls = hookCalls.get(functionNode);
+    if (fnHookCalls === undefined) {
+      return null;
+    }
+    return {
+      key: fnHookCalls.map(call => call.name + '{' + call.key + '}').join('\n'),
+      customHooks: fnHookCalls
+        .filter(call => !isBuiltinHook(call.name))
+        .map(call => call.callee),
+    };
+  }
+
+  function createArgumentsForSignature(node, signature) {
+    const {key, customHooks} = signature;
+    const args = [node, t.stringLiteral(key)];
+    if (customHooks.length > 0) {
+      args.push(t.arrowFunctionExpression([], t.arrayExpression(customHooks)));
+    }
+    return args;
+  }
+
+  let seenForRegistration = new WeakSet();
+  let seenForSignature = new WeakSet();
+  let seenForHookCalls = new WeakSet();
+  let seenForOutro = new WeakSet();
+
   return {
     visitor: {
       ExportDefaultDeclaration(path) {
@@ -171,6 +250,15 @@ export default function(babel) {
           // are currently ignored.
           return;
         }
+
+        // Make sure we're not mutating the same tree twice.
+        // This can happen if another Babel plugin replaces parents.
+        if (seenForRegistration.has(node)) {
+          return;
+        }
+        seenForRegistration.add(node);
+        // Don't mutate the tree above this point.
+
         // This code path handles nested cases like:
         // export default memo(() => {})
         // In those cases it is more plausible people will omit names
@@ -183,6 +271,13 @@ export default function(babel) {
           inferredName,
           declPath,
           (persistentID, targetExpr, targetPath) => {
+            if (targetPath === null) {
+              // For case like:
+              // export default hoc(Foo)
+              // we don't want to wrap Foo inside the call.
+              // Instead we assume it's registered at definition.
+              return;
+            }
             const handle = createRegistration(programPath, persistentID);
             targetPath.replaceWith(
               t.assignmentExpression('=', handle, targetExpr),
@@ -190,7 +285,158 @@ export default function(babel) {
           },
         );
       },
-      FunctionDeclaration(path) {
+      FunctionDeclaration: {
+        enter(path) {
+          const node = path.node;
+          let programPath;
+          let insertAfterPath;
+          switch (path.parent.type) {
+            case 'Program':
+              insertAfterPath = path;
+              programPath = path.parentPath;
+              break;
+            case 'ExportNamedDeclaration':
+              insertAfterPath = path.parentPath;
+              programPath = insertAfterPath.parentPath;
+              break;
+            case 'ExportDefaultDeclaration':
+              insertAfterPath = path.parentPath;
+              programPath = insertAfterPath.parentPath;
+              break;
+            default:
+              return;
+          }
+          const id = node.id;
+          if (id === null) {
+            // We don't currently handle anonymous default exports.
+            return;
+          }
+          const inferredName = id.name;
+          if (!isComponentishName(inferredName)) {
+            return;
+          }
+
+          // Make sure we're not mutating the same tree twice.
+          // This can happen if another Babel plugin replaces parents.
+          if (seenForRegistration.has(node)) {
+            return;
+          }
+          seenForRegistration.add(node);
+          // Don't mutate the tree above this point.
+
+          // export function Named() {}
+          // function Named() {}
+          findInnerComponents(
+            inferredName,
+            path,
+            (persistentID, targetExpr) => {
+              const handle = createRegistration(programPath, persistentID);
+              insertAfterPath.insertAfter(
+                t.expressionStatement(
+                  t.assignmentExpression('=', handle, targetExpr),
+                ),
+              );
+            },
+          );
+        },
+        exit(path) {
+          const node = path.node;
+          const id = node.id;
+          if (id === null) {
+            return;
+          }
+          const signature = getHookCallsSignature(node);
+          if (signature === null) {
+            return;
+          }
+
+          // Make sure we're not mutating the same tree twice.
+          // This can happen if another Babel plugin replaces parents.
+          if (seenForSignature.has(node)) {
+            return;
+          }
+          seenForSignature.add(node);
+          // Don't muatte the tree above this point.
+
+          // Unlike with __register__, this needs to work for nested
+          // declarations too. So we need to search for a path where
+          // we can insert a statement rather than hardcoding it.
+          let insertAfterPath = null;
+          path.find(p => {
+            if (p.parentPath.isBlock()) {
+              insertAfterPath = p;
+              return true;
+            }
+          });
+          if (insertAfterPath === null) {
+            return;
+          }
+
+          insertAfterPath.insertAfter(
+            t.expressionStatement(
+              t.callExpression(
+                t.identifier('__signature__'),
+                createArgumentsForSignature(id, signature),
+              ),
+            ),
+          );
+        },
+      },
+      'ArrowFunctionExpression|FunctionExpression': {
+        exit(path) {
+          const node = path.node;
+          const signature = getHookCallsSignature(node);
+          if (signature === null) {
+            return;
+          }
+
+          // Make sure we're not mutating the same tree twice.
+          // This can happen if another Babel plugin replaces parents.
+          if (seenForSignature.has(node)) {
+            return;
+          }
+          seenForSignature.add(node);
+          // Don't mutate the tree above this point.
+
+          if (path.parent.type === 'VariableDeclarator') {
+            let insertAfterPath = null;
+            path.find(p => {
+              if (p.parentPath.isBlock()) {
+                insertAfterPath = p;
+                return true;
+              }
+            });
+            if (insertAfterPath === null) {
+              return;
+            }
+            // Special case when a function would get an inferred name:
+            // let Foo = () => {}
+            // let Foo = function() {}
+            // We'll add signature it on next line so that
+            // we don't mess up the inferred 'Foo' function name.
+            insertAfterPath.insertAfter(
+              t.expressionStatement(
+                t.callExpression(
+                  t.identifier('__signature__'),
+                  createArgumentsForSignature(path.parent.id, signature),
+                ),
+              ),
+            );
+            // Result: let Foo = () => {}; __signature(Foo, ...);
+          } else {
+            // let Foo = hoc(() => {})
+            path.replaceWith(
+              t.callExpression(
+                t.identifier('__signature__'),
+                createArgumentsForSignature(node, signature),
+              ),
+            );
+            // Result: let Foo = hoc(__signature(() => {}, ...))
+          }
+        },
+      },
+      VariableDeclaration(path) {
+        const node = path.node;
         let programPath;
         let insertAfterPath;
         switch (path.parent.type) {
@@ -209,39 +455,15 @@ export default function(babel) {
           default:
             return;
         }
-        const id = path.node.id;
-        if (id === null) {
-          // We don't currently handle anonymous default exports.
+
+        // Make sure we're not mutating the same tree twice.
+        // This can happen if another Babel plugin replaces parents.
+        if (seenForRegistration.has(node)) {
           return;
         }
-        const inferredName = id.name;
-        if (!isComponentishName(inferredName)) {
-          return;
-        }
-        // export function Named() {}
-        // function Named() {}
-        findInnerComponents(inferredName, path, (persistentID, targetExpr) => {
-          const handle = createRegistration(programPath, persistentID);
-          insertAfterPath.insertAfter(
-            t.expressionStatement(
-              t.assignmentExpression('=', handle, targetExpr),
-            ),
-          );
-        });
-      },
-      VariableDeclaration(path) {
-        let programPath;
-        switch (path.parent.type) {
-          case 'Program':
-            programPath = path.parentPath;
-            break;
-          case 'ExportNamedDeclaration':
-          case 'ExportDefaultDeclaration':
-            programPath = path.parentPath.parentPath;
-            break;
-          default:
-            return;
-        }
+        seenForRegistration.add(node);
+        // Don't mutate the tree above this point.
+
         const declPaths = path.get('declarations');
         if (declPaths.length !== 1) {
           return;
@@ -252,12 +474,70 @@ export default function(babel) {
           inferredName,
           declPath,
           (persistentID, targetExpr, targetPath) => {
+            if (targetPath === null) {
+              // For case like:
+              // export const Something = hoc(Foo)
+              // we don't want to wrap Foo inside the call.
+              // Instead we assume it's registered at definition.
+              return;
+            }
             const handle = createRegistration(programPath, persistentID);
-            targetPath.replaceWith(
-              t.assignmentExpression('=', handle, targetExpr),
-            );
+            if (
+              (targetExpr.type === 'ArrowFunctionExpression' ||
+                targetExpr.type === 'FunctionExpression') &&
+              targetPath.parent.type === 'VariableDeclarator'
+            ) {
+              // Special case when a function would get an inferred name:
+              // let Foo = () => {}
+              // let Foo = function() {}
+              // We'll register it on next line so that
+              // we don't mess up the inferred 'Foo' function name.
+              insertAfterPath.insertAfter(
+                t.expressionStatement(
+                  t.assignmentExpression('=', handle, declPath.node.id),
+                ),
+              );
+              // Result: let Foo = () => {}; _c1 = Foo;
+            } else {
+              // let Foo = hoc(() => {})
+              targetPath.replaceWith(
+                t.assignmentExpression('=', handle, targetExpr),
+              );
+              // Result: let Foo = _c1 = hoc(() => {})
+            }
           },
         );
+      },
+      CallExpression(path) {
+        const node = path.node;
+        const callee = node.callee;
+
+        let name = null;
+        switch (callee.type) {
+          case 'Identifier':
+            name = callee.name;
+            break;
+          case 'MemberExpression':
+            name = callee.property.name;
+            break;
+        }
+        if (name === null || !/^use[A-Z]/.test(name)) {
+          return;
+        }
+
+        // Make sure we're not recording the same calls twice.
+        // This can happen if another Babel plugin replaces parents.
+        if (seenForHookCalls.has(node)) {
+          return;
+        }
+        seenForHookCalls.add(node);
+        // Don't mutate the tree above this point.
+
+        const fn = path.scope.getFunctionParent();
+        if (fn === null) {
+          return;
+        }
+        recordHookCall(fn.block, path, name);
       },
       Program: {
         exit(path) {
@@ -265,16 +545,28 @@ export default function(babel) {
           if (registrations === undefined) {
             return;
           }
+
+          // Make sure we're not mutating the same tree twice.
+          // This can happen if another Babel plugin replaces parents.
+          const node = path.node;
+          if (seenForOutro.has(node)) {
+            return;
+          }
+          seenForOutro.add(node);
+          // Don't mutate the tree above this point.
+
           registrationsByProgramPath.delete(path);
           const declarators = [];
           path.pushContainer('body', t.variableDeclaration('var', declarators));
           registrations.forEach(({handle, persistentID}) => {
             path.pushContainer(
               'body',
-              buildRegistrationCall({
-                HANDLE: handle,
-                PERSISTENT_ID: t.stringLiteral(persistentID),
-              }),
+              t.expressionStatement(
+                t.callExpression(t.identifier('__register__'), [
+                  handle,
+                  t.stringLiteral(persistentID),
+                ]),
+              ),
             );
             declarators.push(t.variableDeclarator(handle));
           });
